requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
xgboost>=2.0.3
lightgbm>=4.3.0
catboost>=1.2.2
pandas>=2.2.0
numpy>=1.26.4,<2.0.0
scikit-learn>=1.3.2
matplotlib>=3.8.2
python-binance>=1.0.19
python-dotenv>=1.0.0
pytest>=8.0.0
httpx>=0.26.0
ta>=0.11.0
ccxt>=4.0.0
optuna>=3.5.0
aiohttp>=3.9.1
python-multipart>=0.0.6
apscheduler>=3.10.4
pydantic>=2.5.3
pydantic-settings>=2.1.0
torch>=2.0.0
prometheus_client>=0.19.0
psutil>=5.9.0
redis>=5.0.0
aiosqlite>=0.19.0
gymnasium>=1.0.0
